Updated formatting, added codejam snippet

+++ power.cpp
@@ -1,7 +1,7 @@
 #include "bits/stdc++.h"
 using namespace std;
 
-const int mod = 1e9 + 7;
+const int mod = (int) 1e9 + 7;
 
 int getpow (int a, int b) {
 	if (b == 0) return 1;
@@ -9,8 +9,7 @@ int getpow (int a, int b) {
 	if (b % 2 == 0) {
 		int t = getpow(a, b / 2);
 		return (t * t) % mod;
-	}
-	else {
+	} else {
 		int t = getpow(a, (b - 1) / 2);
 		return (((a * t) % mod) * t) % mod;
 	}
